requirements.txt
aicspylibczi>=0.0.0
asciitree>=0.3.3
bfio>=0.0.0
bioformats_jar>=0.0.0
bioio-base>=0.0.0
bioio-bioformats==1.1.0
bioio-czi==2.1.0
bioio-imageio==1.1.0
bioio-lif==1.1.0
bioio-nd2==1.1.0
bioio-ome-tiff-fork-by-bugra==0.0.1b2
bioio-tifffile-fork-by-bugra>=0.0.1b2
cmake==4.0.2
dask>=2024.12.1
dask-jobqueue>=0.0.0
distributed>=2024.12.1
elementpath==5.0.1
fasteners==0.19
fire>=0.0.0
imageio==2.27.0
imageio-ffmpeg==0.6.0
install-jdk
natsort>=0.0.0
nd2>=0.0.0
numpy>=1.24
pydantic>=2.11.7
pylibczirw>=0.0.0
readlif==0.6.5
s3fs>=0.0.0
scipy>=1.8
tensorstore>=0.0.0
tifffile>=2025.5.21
validators==0.35.0
xarray>=0.0.0
xmlschema>=0.0.0
xmltodict==0.14.2
zarr>=3.0
zstandard>=0.0.0
aiofiles>=24.1.0
blosc2>=3.7.1
fastapi>=0.116.1
lz4>=4.4.4
psutil>=7.0.0
rich>=14.1.0
uvicorn>=0.35.0
websockets>=15.0.1
h5py